fixed new night stalling bug
night count was not incremented properly when a night ends. it was originally incremented only when the player presses continue, but then that would mean going back to the main menu would restart the count. the fix is to increment night count after the continue screen displays its text. new game is also set to false once the continue screen is reached