restructured storage system to support multiple
created a storage controller singleton for easy access of hold and cooldown meters and to keep track of the current storage being interacted with. storage now has independent cooldown times that will update the cooldown meter value accordingly when it is being triggered. hold meter is now handled completely in storage controller