rewrote meter script to be more autonomous
rewrote all of meter so that pausing and resuming only relies on the game manager's current state rather than manual calls to its functions. simplified the logic for setting a new multiplier and changing the meter value by a specified amount. now meter logic is more readable and easier to follow