fixed qte inputs to be more responsive
original method took in horizontal and vertical inputs which is checked every update. but horizontal and vertical inputs can be 1/-1 for more than one frame even on one key press, so a time delay variable was initialized to separate each key press which causes a correct keypress to be missed. modified this method to check for GetKeyDown on AWSD