added "night vision", adjusted light ellipse+specs
rooms are not lit by default by a dim global light so its not entirely black when the candlelight goes out to simulate night vision. light eclipse around the player is also shortened so that the candle in the sideviews will be closer to the mc. light radius is also reduced