qte ui and scene resets, fixed singleton null ref
if mc died on the stairs the qte ui would still show up when returning to the game, which is now fixed to not display on reset. scene now resets to the attic regardless of where the mc died. bug: when pause is pressed and we wait for the end of the night, the continue screen pops up and displays the wrong stats bc of a null reference to the continue script. somehow it was being duplicated which triggered the default singleton functionality to delete the existing instance